+++ src/helpers/notificationHelper.js
@@ -1,0 +1,5 @@
+import { generateTitle } from './titleHelper';
+
+export function displayNotification(document, originalTitle, replaceTitle, value) {
+  document.title = generateTitle(value, originalTitle, replaceTitle);
+}